% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucGARP.R
\name{aucGARP}
\alias{aucGARP}
\title{Area Under the Receiver Operating Characteristic (ROC) curve (AUC)}
\usage{
aucGARP(n, x, points)
}
\arguments{
\item{n}{a numeric value specifying the number of models in the best subset outputted by DesktopGARP}

\item{x}{a raster object of the summated raster of the best models output by DesktopGARP}

\item{points}{a spatial object of presence data to use for testing locations}
}
\value{
Plots the Receiver Operating Characteristic (ROC) curve and returns a list containing:
\itemize{
\item \code{AUC} the total area under the ROC curve
\item \code{Wilcoxon} the Wilcoxon test statistic; essentially equal to the AUC
\item \code{Standard.Error} the standard error of the AUC value
\item \code{Z.score} the z-score associated with the AUC value
}
}
\description{
\code{aucGARP} Calculates the area under the ROC curve (AUC), which represents the accuracy of the subset of the best models produced by GARP. AUC evaluates the predictive performance of the
best models by relating the model sensitivity (true positive rate) to 1-specificity (true negative rate), and can be described as the probability that any given cell is correctly predicted
as present or absent. An AUC of 0.5 is that predicted at random, while an AUC of 1 represents a perfect prediction
}
\details{
When calculating AUC for discrete cutpoints (\code{n}), represented by the number of models that agree on a predicted presence location, the AUC of the entire curve
is essentially equal to the Wilcoxon test statistic.
The raster object (\code{x}) should be a raster representing the number of models that agree on a predicted presence location per pixel and that outtput by \code{\link{sumRasters}}.
The shapefile \code{points} should presence locations that were not used by GARP for model training and those output by \code{\link{splitData}}.
}
\examples{
  set.seed(0)
  library(raster)
  r   <- raster(ncols = 100, nrows = 100)
  r[] <- rbinom(5, 10, 0.3)
  hs  <- data.frame("Latitude" = c(-89, 72, 63, 42, 54), "Longitude" = c(-12, 13, 24, 26, 87), "Species" = rep("Homo_sapiens", 5))
  aucGARP(n = 10, x = r, points = SpatialPoints(hs[,1:2]))

}
\references{
Cortes, C. and Mohri, M. (2004) Confidence Intervals for the Area Under the ROC curve. \emph{Advances in Neural Information Processing Systems}. \bold{6}.
}
\seealso{
\code{\link{zAUC}}, \code{\link{seAUC}}, \code{\link{wGARP}}, \code{\link{plotROC}}
}
