% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucRatio.R
\name{aucRatio}
\alias{aucRatio}
\title{Ratio of Area Under the Receiver Operating Characteristic (ROC) curve (AUC)}
\usage{
aucRatio(n, x, points, E)
}
\arguments{
\item{n}{a numeric value specifying the number of models in the best subset outputted by GARP}

\item{x}{a raster object of the summated raster of the best models output by GARP}

\item{points}{a spatial object of presence data to use for testing locations}

\item{E}{the amount of error admissible along the true-positive axis (less than 1.0)}
}
\value{
Plots the modified Receiver Operating Characteristic (ROC) curve and returns a list containing:
\itemize{
\item \code{Modified.AUC} the total area under the modified ROC curve
\item \code{Modified.Reference} the area under the modified reference curve
\item \code{AUC.ratio} the ratio of the modified AUC to the modified reference value
}
}
\description{
\code{aucRatio} Calculates the ratio of the area under the ROC curve (AUC) as suggested by Peterson et al. (2008). AUC represents the accuracy of the subset of the best models produced by GARP. AUC evaluates the predictive performance of the
best models by relating the model sensitivity (true positive rate) to 1-specificity (true negative rate), and can be described as the probability that any given cell is correctly predicted
as present or absent. An AUC of 0.5 is that predicted at random, while an AUC of 1 represents a perfect prediction
}
\details{
The raster object (\code{x}) should be a raster representing the number of models that agree on a predicted presence location per pixel and that outtput by \code{\link{sumRasters}}.
The shapefile \code{points} should presence locations that were not used by GARP for model training and those output by \code{\link{splitData}}.
}
\examples{
  set.seed(0)
  library(raster)
  r   <- raster(ncols = 100, nrows = 100)
  r[] <- rbinom(5, 10, 0.3)
  hs  <- data.frame("Latitude" = c(-89, 72, 63, 42, 54), "Longitude" = c(-12, 13, 24, 26, 87), "Species" = rep("Homo_sapiens", 5))
  aucRatio(n = 10, x = r, points = SpatialPoints(hs[,1:2]), E = 0.20)

}
\references{
Peterson, A.T., M. Papes, J. Soberon (2007) Rethinking receiver operating characteristic analysis applications in ecological niche modeling. \emph{Ecological Modelling}. \bold{213(1):63-72}.
}
\seealso{
\code{\link{aucGARP}}, \code{\link{zAUC}}, \code{\link{seAUC}}, \code{\link{wGARP}}, \code{\link{plotROC}}
}
