% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carolinas_boundary-data.R
\docType{data}
\name{carolinas_boundary}
\alias{carolinas_boundary}
\title{State boundary of North and South Carolina, United States}
\format{
An object of class \code{"SpatialPolygonsDataFrame"}
}
\usage{
data(carolinas_boundary)
}
\description{
A spatial polygon of the state boundaries of North and South Carolina, United States.
}
\examples{
library(sp)
data(carolinas_boundary)
plot(carolinas_boundary)
}
\keyword{datasets}
