% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\title{Returns centroid coordinates of raster cells containing presence locations}
\usage{
centroid(x, points, xy = NULL, species, output.path = NULL)
}
\arguments{
\item{x}{an object of class raster that represents the spatial resolution and extent of raster inputted into GARP}

\item{points}{a data.frame or SpatialPointsDataFrame object containing latitude and longitude values of presence locations}

\item{xy}{a vector object of latitude and longitude coordinates from \code{points} when \code{points} is class data.frame. Not needed if \code{points} is a SpatialPointsDataFrame object}

\item{species}{the vector of species names from the \code{points} data.frame or SpatialPointsDataFrame object}
}
\value{
A SpatialPointsDataFrame object containing "Latitude", "Longitude", and "Species" vectors.
}
\description{
\code{centroid} The desktop version of GARP records the instance of a presence location in a grid cell, and therefore does not distinguish between multiple locations within a single cell. This function
returns the centroid location of all grid cells that contain a presence location, which circumvents multiple locations. It is useful when input rasters cells are large or
when sampling was unevenly clustered. This function uses the cell size and extent of user-defined raster to generate centroid coordinates of all cells that contain presence locations.
}
\details{
\code{x} should be a raster that represents the spatial resolution and extent of all environmental layers that will be inputted into GARP. Recommended to use a raster output by \code{\link{rasterPrep}}.
}
\examples{
set.seed(0)
library(raster)
r   <- raster(ncols = 100, nrows = 100)
r[] <- rbinom(5, 10, 0.3)
hs  <- data.frame("Latitude" = c(-89, 72, 63, 42, 54),"Longitude" = c(-12, 13, 24, 26, 87), "Species" = rep("Homo_sapiens", 5))
centroid(x = r, points = hs, xy = cbind(hs$Latitude, hs$Longitude), species = hs$Species)

}
