% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confuseMatrix.R
\name{confuseMatrix}
\alias{confuseMatrix}
\title{Calculates the confusion matrix}
\usage{
confuseMatrix(n, x, points)
}
\arguments{
\item{n}{a numeric value specifying the number of models in the best subset output from GARP}

\item{x}{a raster object of the summated raster of the best models output from GARP}

\item{points}{a spatial object of presence data to use for testing locations}
}
\value{
A data.frame object containing:
\itemize{
\item \code{a} True positive
\item \code{b} False negative
\item \code{c} False positive
\item \code{d} True negative
\item \code{sensitivity} the proportion of occupied pixels correctly predicted as present
\item \code{specificity} the proportion of non-occupied pixels correctly predicted as absent
\item \code{1-specificity} the proportion of non-occupied pixels incorrectly predicted as present
}
}
\description{
\code{confuseMatrix} Returns the confusion matrix containing the sensitivity and specificity of the best model subset of a GARP run.
}
\details{
For discrete cutpoints (\code{n}), represents the number of models that agree on a predicted presence location.

The raster object (\code{x}) should be a raster representing the number of models that agree on a predicted presence location per pixel and that outtput by \code{\link{sumRasters}}.

The shapefile \code{points} should presence locations that were not used by GARP for model training and those output by \code{\link{splitData}}
}
\examples{
  set.seed(0)
  library(raster)
  r   <- raster(ncols = 100, nrows = 100)
  r[] <- rbinom(5, 10, 0.3)
  hs  <- data.frame("Latitude" = c(-89, 72, 63, 42, 54), "Longitude" = c(-12, 13, 24, 26, 87), "Species" = rep("Homo_sapiens", 5))
  confuseMatrix(n = 10, x = r, points = SpatialPoints(hs[,1:2]))

}
\seealso{
\code{\link{aucGARP}}
}
