% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_layers-data.R
\docType{data}
\name{env_layers}
\alias{env_layers}
\title{Environmental layers from WorldClim}
\format{
An object of class \code{"RasterStack"}
}
\source{
\href{https://www.worldclim.org/data/bioclim.html}{WorldClim}
}
\usage{
data(env_layers)
}
\description{
A RasterStack of three environmental grids (bio_1: annual mean temperature, bio_12: annual precipitation, and alt: altitude)
}
\examples{
data(env_layers)
}
\keyword{datasets}
