% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRules.R
\name{extractRules}
\alias{extractRules}
\title{Pull rulesets created by GARP}
\usage{
extractRules(file, table, pathA, pathB, project)
}
\arguments{
\item{file}{file name or file path of the ".txt" file of rulesets generated by GARP}

\item{table}{a data.frame object produced by \code{\link{findModRules}} that contains model numbers (ruleset number) and dominant rule number}

\item{pathA}{file path to the directory containing environmental grids (must be in ESRI ASCII raster format)}

\item{pathB}{file path to the directory for ruleset grids produced by GARP}

\item{project}{logical; indicates if the GARP results are projected to the landscape}
}
\value{
Returns a data.frame containing the minimum and maximum values for the environmental variables in each rule, the type of rules, the model numbers (ruleset number) and rule numbers.
}
\description{
\code{extractRules} Extract the dominant range rules and logit rules for the best subset models
}
\details{
Include both the path and the full file name in \code{file}, otherwise, set the directory to the rulesets file before running the function.
}
\examples{
\dontrun{
  rangelogit <- extractRules("C:/GARP/RuleSets.txt", table = modelRule,
  pathA = "C:/GARP/resampled/", pathB = "C:/GARP/runs/", project = FALSE)
}

}
\seealso{
\code{\link{findModRules}}
}
