% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findModRules.R
\name{findModRules}
\alias{findModRules}
\title{Pull rulesets created by GARP}
\usage{
findModRules(pathA, pathB)
}
\arguments{
\item{pathA}{file path to the directory for ruleset grids produced by GARP}

\item{pathB}{a vector object that contains the directories of the best subsets for each GARP experiment}
}
\value{
Returns a data.frame that includes the model numbers (ruleset number) and dominant rule numbers
}
\description{
\code{findModRules} Gets the dominant rules produced by GARP which cover more than 90\% of the landscape for the best subset models
}
\examples{
\dontrun{
modelRule <- findModRules(pathA = "C:/GARP/runs/", pathB = "C:/GARP/BestSubsets/")
}

}
\seealso{
\code{\link{extractRules}}
}
