% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMedMinMax.R
\name{getMedMinMax}
\alias{getMedMinMax}
\title{Pull rulesets created by GARP}
\usage{
getMedMinMax(pathA, pathB)
}
\arguments{
\item{pathA}{a vector object that contains the full file names for the extracted rule .csv files created by \code{\link{extractRules}}}

\item{pathB}{a file path of the directory containing environmental layers used in GARP (must be in ESRI ASCII raster format)}
}
\value{
Returns a data.frame that contains the range, and median of the minimum and maximum values for each environment variable used in GARP.
}
\description{
\code{getMedMinMax} Returns the minimum and maximum median values for variables in dominant rules from a given rule-set
}
\examples{
\dontrun{
MedMinMax <- getMedMinMax(pathA = "C:/GARP/RangeLogit.csv", pathB = "C:/GARP/resampled/")
}

}
\seealso{
\code{\link{extractRules}}
}
