% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_boundary-data.R
\docType{data}
\name{nc_boundary}
\alias{nc_boundary}
\title{State boundary of North Carolina, United States}
\format{
An object of class \code{"SpatialPolygonsDataFrame"}
}
\usage{
data(nc_boundary)
}
\description{
A spatial polygon of the state boundarie of North Carolina, United States.
}
\examples{
library(sp)
data(nc_boundary)
plot(nc_boundary)
}
\keyword{datasets}
