% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omissionGARP.R
\name{omissionGARP}
\alias{omissionGARP}
\title{Calculates omission metrics of model outputs}
\usage{
omissionGARP(n, x, points)
}
\arguments{
\item{n}{a numeric value specifying the number of models in the best subset output from GARP}

\item{x}{a raster object of the summated raster of the best models output from GARP}

\item{points}{a spatial object of presence data to use for testing locations}
}
\value{
A list object containing:
\itemize{
\item \code{Average.Omission} the average omission of the model subset
\item \code{Total.Omission} the total omission of the model subset
}
}
\description{
\code{omissionGARP} Calculates the omission, or the true occurrence points that are not predicted as present, of the best models output from GARP.
}
\details{
The discrete cutpoints (\code{n}) represent the number of models that agree on a predicted presence location.

The raster object (\code{x}) should be a raster representing the number of models that agree on a predicted presence location per pixel and that outtput by \code{\link{sumRasters}}.

The shapefile \code{points} should presence locations that were not used by GARP for model training and those output by \code{\link{splitData}}.
}
\examples{
  set.seed(0)
  library(raster)
  r   <- raster(ncols = 100, nrows = 100)
  r[] <- rbinom(5, 10, 0.3)
  hs  <- data.frame("Latitude" = c(-89, 72, 63, 42, 54), "Longitude" = c(-12, 13, 24, 26, 87), "Species" = rep("Homo_sapiens", 5))
  omissionGARP(n = 10, x = r, points = SpatialPoints(hs[,1:2]))

}
\seealso{
\code{\link{commissionGARP}}
}
