% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{prevalence}
\alias{prevalence}
\title{Calculate the prevalence of each environmental covariate in the rulesets}
\usage{
prevalence(pathA, pathB, num)
}
\arguments{
\item{pathA}{a vector object that contains the full file names for the extracted rule .csv files created by \code{\link{extractRules}}}

\item{pathB}{the primary directory for ruleset grids}

\item{num}{numeric; the total number of dominant presence rules; can be generated from \code{\link{totPresRules}}}
}
\value{
A table recording the prevalence of each environmental covariate.
}
\description{
\code{prevalence} Calculate the prevalence of each environmental covariate in the ruleset file
}
\details{
Ruleset grids called by the \code{pathB} path must be in ESRI ASCII format.
}
