% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPrep.R
\name{rasterPrep}
\alias{rasterPrep}
\title{Prepares input rasters for GARP}
\usage{
rasterPrep(
  file.path,
  res.raster = NULL,
  cells = NULL,
  mask = NULL,
  rescale = FALSE,
  output.path
)
}
\arguments{
\item{file.path}{location of rasters to resample and crop}

\item{res.raster}{optional; a raster object to resample and crop all other rasters to}

\item{cells}{optional; a numeric value representing the number of cells per row to resample rasters}

\item{mask}{optional; a spatial object specifying extent to crop rasters}

\item{rescale}{logical; should each raster be centered on the mean and scaled by the standard deviation?}

\item{output.path}{a file path specifying where the output rasters should be saved}
}
\value{
For each raster within \code{file.path}, a new ACII file, resampled to the spatial resolution of the sample raster and spatial extent of the sample raster (\code{res.raster}) or \code{mask} file,
will be returned to \code{output.path} with the same name and \code{"_resample"} appended to the file name.
}
\description{
\code{rasterPrep} The desktop version of GARP requires all input environmental layers to be of the same spatial resolution and extent and saved as an ".asc". This function resamples input rasters to the same cell size and extent of a sample raster.
User can define sample raster in \code{res.raster} or define the number of cells in each row of the raster with \code{cells}. If both \code{res.raster} and \code{cell} parameters are \code{NULL},
raster with the largest cell size will be automatically defined as the sample raster. The user can also crop rasters to a mask of the study area with \code{mask = "mask filename"}.
}
