% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescales environmental variables from 0 to 1}
\usage{
rescale(file)
}
\arguments{
\item{file}{a data.frame of the original median of the minimum and maximum values for each environmental variable produced by \code{\link{getMedMinMax}}}
}
\value{
Returns a data.frame containing the rescaled range, and the median of the minimum and maximum values for each environmental variable.
}
\description{
\code{rescale} Rescales the the minimum and maximum median values for each environment variable used in dominant rules across a GARP Best Subsets Procedure to be between 0 to 1. This will allow the user to evalaute the range of values across variables regardless of units of measure
}
\seealso{
\code{\link{getMedMinMax}}
}
