% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seAUC.R
\name{seAUC}
\alias{seAUC}
\title{Calculates the standard error of the AUC}
\usage{
seAUC(n, x, points)
}
\arguments{
\item{n}{a numeric value specifying the number of models in the best subset output from GARP}

\item{x}{a raster object of the summated raster of the best models output from GARP}

\item{points}{a spatial object of presence data to use for testing locations}
}
\value{
Returns the standard error of the AUC.
}
\description{
\code{seAUC} Returns the standard error of the AUC of the best model subset output from GARP.
}
\details{
For discrete cutpoints (\code{n}), represents the number of models that agree on a predicted presence location.

The raster object (\code{x}) should be a raster representing the number of models that agree on a predicted presence location per pixel and that outtput by \code{\link{sumRasters}}.

The shapefile \code{points} should presence locations that were not used by DesktopGARP for model training and those output by \code{\link{splitData}}.
}
\examples{
  set.seed(0)
  library(raster)
  r   <- raster(ncols = 100, nrows = 100)
  r[] <- rbinom(5, 10, 0.3)
  hs  <- data.frame("Latitude" = c(-89, 72, 63, 42, 54), "Longitude" = c(-12, 13, 24, 26, 87), "Species" = rep("Homo_sapiens", 5))
  seAUC(n = 10, x = r, points = SpatialPoints(hs[,1:2]))

}
\seealso{
\code{\link{aucGARP}}
}
