% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumRasters.R
\name{sumRasters}
\alias{sumRasters}
\title{Sums rasters of the best subset models output from GARP}
\usage{
sumRasters(
  file.path,
  output = FALSE,
  output.path = NULL,
  format = NULL,
  over.write = FALSE
)
}
\arguments{
\item{file.path}{path to the location of output rasters to sum (usually found in "GARPruns" folder output from GARP)}

\item{output}{logical. Should the summated raster be saved to file?}

\item{output.path}{a file path specifying where the output raster should be saved, including file name and type (e.g. "C:/Users/sumraster1.asc"). Can be of type ".grd", ".asc", ".tif", etc. See \href{https://cran.r-project.org/web/packages/raster/raster.pdf}{writeFormats} for supported types}

\item{format}{the format type of the raster to be saved. Can be of type "raster", "ascii", "GTiff", etc. See \href{https://cran.r-project.org/web/packages/raster/raster.pdf}{writeFormats} for supported types}

\item{over.write}{logical. Should the raster be overwritten?}
}
\value{
A single raster with cell values that represent the number of models that agree on a predicted presence location.
}
\description{
\code{sumRasters} GARP outputs binary presence/absence grids for the sample area and selects the best models based on user-defined qualifications. This function
sums the best models into a single raster that represents the number of models that agree on a predicted presence location.
}
