% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks_layers-data.R
\docType{data}
\name{tasks_layers}
\alias{tasks_layers}
\title{ESRI output grids from a single GARP experiment}
\format{
An object of class \code{"RasterStack"}
}
\usage{
data(tasks_layers)
}
\description{
A RasterStack of the 10 individual binary presence/absence ESRI grids from a single GARP experiment for white-tailed deer and 3 bioclimatic variables (altitude, annual mean temperature, annual precipitation)
}
\examples{
library(raster)
data(tasks_layers)
}
\keyword{datasets}
