% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totPresRules.R
\name{totPresRules}
\alias{totPresRules}
\title{Get the total number of dominant presence rules in the ruleset file}
\usage{
totPresRules(file, table, pathA, pathB, project)
}
\arguments{
\item{file}{file name or file path of the ".txt" file of rulesets generated by GARP}

\item{table}{a data.frame object produced by \code{\link{findModRules}} that contains model numbers (ruleset number) and dominant rule number}

\item{pathA}{file path to the directory containing environmental grids (must be in ESRI ASCII raster format)}

\item{pathB}{file path to the directory for ruleset grids produced by GARP}

\item{project}{logical; indicates if the GARP results are projected to the landscape}
}
\value{
A number of the total number of dominant presence rules.
}
\description{
\code{totPresRules} Get the total number of dominant presence rules in the ruleset file
}
\details{
While specifying the \code{file} element, include both the path and the full file name,
otherwise, set the directory to the rulesets file before running the function.
}
\examples{
\dontrun{
PreNo <- totPresRules(file = "C:/GARP/RuleSets.txt", table = modelRule,
pathA = "C:/GARP/resampled/", pathB = "C:/GARP/runs/", project = FALSE)
}

}
