% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimportIdx.R
\name{unimportIdx}
\alias{unimportIdx}
\title{Calculates the unimportance index of each environmental covariate for variable selection}
\usage{
unimportIdx(table1, table2)
}
\arguments{
\item{table1}{a data.frame recording the prevalence of each environmental covariate; can be produced by \code{\link{prevalence}}}

\item{table2}{a data.frame with the rescaled median range for each environmental variable; can be produced by \code{\link{rescale}}}
}
\value{
A table recording the normalized unimportance index for each environmental covariate
}
\description{
\code{unimportIdx} Calculates the unimportance index of each environmental covariate for variable selection
}
