% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtdeer_df-data.R
\docType{data}
\name{wtdeer_df}
\alias{wtdeer_df}
\title{Example white-tailed deer GPS locations}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{www.gbif.org}{GBIF.org}
}
\usage{
data(wtdeer_df)
}
\description{
Latitude and longitude data from white-tailed (\emph{Odocoileus virginianus}) in North and South Carolina, United States.
}
\examples{
data(wtdeer_df)
head(wtdeer_df)
}
\keyword{datasets}
