% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtdeer_locations-data.R
\docType{data}
\name{wtdeer_locations}
\alias{wtdeer_locations}
\title{Spatial data of white-tailed deer GPS locations}
\format{
An object of class \code{"SpatialPointsDataFrame"}
}
\usage{
data(wtdeer_locations)
}
\description{
Spatial layer of GPS locations from white-tailed (\emph{Odocoileus virginianus}) in North and South Carolina, United States.
}
\examples{
library(sp)
data(wtdeer_locations)
plot(wtdeer_locations)
}
\keyword{datasets}
