% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRange.R
\name{plotRange}
\alias{plotRange}
\title{Plots the rescaled median ranges for environment variables}
\usage{
plotRange(file, colour, n, xlabel, ylabel)
}
\arguments{
\item{file}{a data.frame of the rescaled range, and the median of the minimum and maximum values for each environmental variables used in DesktopGARP. Produced by \code{\link{rescale}}}

\item{colour}{a character string designating the color of the bars of the plot}

\item{n}{numeric; specifies the width of the bars}

\item{xlabel}{label for the x-axis}

\item{ylabel}{label for the y-axis}
}
\value{
Returns a plot of the rescaled median ranges for environment variables used in DesktopGARP.
}
\description{
\code{plotRange} Plots the rescaled median ranges for environment variables used across dominant rules from a DesktopGARP experiment
}
